+++ mobile/android/geckoview/src/main/java/org/mozilla/geckoview/WebExtension.java
@@ -1630,9 +1630,13 @@ public class WebExtension {
     /** One of {@link ErrorCodes} that provides more information about this exception. */
     public final @Codes int code;
 
+    /** An optional name of the extension that caused the exception. */
+    public final @Nullable String extensionName;
+
     /** For testing */
     protected InstallException() {
       this.code = ErrorCodes.ERROR_NETWORK_FAILURE;
+      this.extensionName = null;
     }
 
     @Override
@@ -1640,8 +1644,14 @@ public class WebExtension {
       return "InstallException: " + code;
     }
 
+    /* package */ InstallException(final @Codes int code, final @Nullable String extensionName) {
+      this.code = code;
+      this.extensionName = extensionName;
+    }
+
     /* package */ InstallException(final @Codes int code) {
       this.code = code;
+      this.extensionName = null;
     }
   }